Clean up CSS, move wrapper etc into component library
