+++ src/components/Skills.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import styled from 'styled-components/macro'
-import { Heading } from './ComponentLibrary'
+import { Heading, Wrapper } from './ComponentLibrary'
 import PageSection from './PageSection'
-import Wrapper from './Wrapper'
 
 const SkillsList = styled.ul`
     font-family: "Roboto", sans-serif;
