+++ src/components/Thoughts.js
@@ -1,7 +1,6 @@
 import React from 'react'
-import { Heading, Subheading } from './ComponentLibrary'
+import { Heading, Subheading, HideOnMobile, Wrapper } from './ComponentLibrary'
 import PageSection from './PageSection'
-import Wrapper from './Wrapper'
 import Blogpost from './Blogpost'
 import blogpostsData from '../blogposts.json'
 
@@ -9,7 +8,8 @@ const Thoughts = () => {
   return (
     <PageSection>
       <Wrapper>
-        <Heading>My thoughts about code</Heading>
+        <Heading>My thoughts<HideOnMobile> about code</HideOnMobile>
+        </Heading>
         {blogpostsData.map((blogpost) => (
           <Blogpost key={blogpost.title} title={blogpost.title} />
         ))}
