--- src/components/Wrapper.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components/macro'
-
-const Wrapper = styled.div`
-  width: 90%;
-  max-width: 800px;
-  margin: 0 auto;
-`
-
-export default Wrapper
